pyproject.toml
[project]
name = "retail-analytics-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "dspy-ai>=2.4.0",
    "langchain-core>=0.2.0",
    "langgraph>=0.1.0",
    "nltk>=3.9.2",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pydantic>=2.0.0",
    "rank-bm25>=0.2.2",
    "rich>=13.7.0",
    "scikit-learn>=1.3.0",
]
